requirements.txt
Django>=4.2,<4.2.1
psycopg2==2.9.6
django-rest-framework
shortuuid